+++ test/tesla/middleware/meta_logger_test.exs
@@ -8,61 +8,82 @@ defmodule Tesla.Middleware.MetaLoggerTest do
   defmodule FakeClient do
     use Tesla
 
-    plug(Subject,
+    plug Subject,
       filter_headers: ["authorization"],
       filter_query_params: [:username],
       filter_body: [{~r/"email":".*?"/, ~s("email":"[FILTERED]")}],
       log_level: :debug,
       log_tag: Subject
-    )
+
+    @response_headers [
+      {"content-type", "text/plain"},
+      {"authorization", "somesecretthatshouldntbelogged"}
+    ]
 
     adapter(fn env ->
-      env =
-        Tesla.put_headers(env, [
-          {"content-type", "text/plain"},
-          {"authorization", "somesecretthatshouldntbelogged"}
-        ])
-
       case env.url do
         "/connection-error" ->
           {:error, :econnrefused}
 
-        "/error" ->
-          {:ok, %{env | status: 404, body: "error"}}
-
-        "/redirect" ->
-          {:ok, %{env | status: 301, body: "moved"}}
-
         "/ok" ->
-          {:ok, %{env | status: 200, body: "ok"}}
+          {:ok, FakeClient.put_response(env, 200, "response body ok")}
+
+        "/huge-response" ->
+          {:ok,
+           FakeClient.put_response(
+             env,
+             200,
+             String.duplicate("a", 100) <> String.duplicate("b", 100)
+           )}
 
         "/json" ->
-          {:ok, %{env | status: 200, body: ~s({"email":"foo@bar.baz","response":"value"})}}
+          {:ok, FakeClient.put_response(env, 200, ~s({"email":"foo@bar.baz","response":"value"}))}
+
+        "/redirect" ->
+          {:ok, FakeClient.put_response(env, 301, "response body moved")}
+
+        "/error" ->
+          {:ok, FakeClient.put_response(env, 404, "response body error")}
       end
     end)
+
+    def put_response(env, status, body),
+      do: %{env | status: status, headers: @response_headers, body: body}
   end
 
   describe "call/3" do
-    test "logs the request and response" do
-      logs = capture_log(fn -> FakeClient.get("/ok", query: [page: 1, username: "test_user"]) end)
+    test "logs the request and the response" do
+      logs =
+        capture_log(fn ->
+          FakeClient.get("/ok",
+            headers: [{"accept", "text/plain"}],
+            query: [page: 1, username: "test_user"]
+          )
+        end)
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] GET /ok [page: 1, username: "[FILTERED]"] [])
+               "[debug] [#{inspect(Subject)}] GET /ok?page=1&username=[FILTERED] " <>
+                 ~s([{"accept", "text/plain"}])
 
       assert logs =~
                "[debug] [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[debug] [#{inspect(Subject)}] ok"
+      assert logs =~ "[debug] [#{inspect(Subject)}] response body ok"
     end
 
-    test "when body is given logs the request and response" do
-      logs = capture_log(fn -> FakeClient.post("/json", ~s({"email":"foo@bar.baz"})) end)
+    test "when the request body is given, logs the request and the response" do
+      logs =
+        capture_log(fn ->
+          FakeClient.post("/json", ~s({"email":"foo@bar.baz","some":"body"}),
+            headers: [{"accept", "application/json"}]
+          )
+        end)
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] POST /json [] [])
+               ~s([debug] [#{inspect(Subject)}] POST /json [{"accept", "application/json"}])
 
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]"})
+      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","some":"body"})
 
       assert logs =~
                "[debug] [#{inspect(Subject)}] 200 " <>
@@ -71,139 +92,170 @@ defmodule Tesla.Middleware.MetaLoggerTest do
       assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","response":"value"})
     end
 
-    test "when log level is given, logs the message with given level" do
-      logs = capture_log(fn -> FakeClient.post("/ok", %{}, opts: [log_level: :info]) end)
+    test "when query string and request body are given, logs the request and the response" do
+      logs =
+        capture_log(fn ->
+          FakeClient.post("/json", ~s({"email":"foo@bar.baz","some":"body"}),
+            headers: [{"accept", "application/json"}],
+            query: [page: 1, username: "test_user"]
+          )
+        end)
 
-      assert logs =~ ~s([info]  [#{inspect(Subject)}] POST /ok [] [])
+      assert logs =~
+               "[debug] [#{inspect(Subject)}] " <>
+                 ~s(POST /json?page=1&username=[FILTERED] [{"accept", "application/json"}])
+
+      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","some":"body"})
+
+      assert logs =~
+               "[debug] [#{inspect(Subject)}] 200 " <>
+                 ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
+
+      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","response":"value"})
+    end
+
+    test "when the log tag is given as a string, logs using the log tag without inspecting" do
+      logs = capture_log(fn -> FakeClient.get("/ok", opts: [log_tag: "MOOI"]) end)
+
+      assert logs =~ "[debug] [MOOI] GET /ok"
+
+      assert logs =~
+               "[debug] [MOOI] 200 " <>
+                 ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
+
+      assert logs =~ "[debug] [MOOI] response body ok"
+    end
+
+    test "when the log tag is given as not a string, logs using the log tag inspecting" do
+      logs = capture_log(fn -> FakeClient.get("/ok", opts: [log_tag: FakeClient]) end)
+
+      assert logs =~ "[debug] [#{inspect(FakeClient)}] GET /ok []"
+
+      assert logs =~
+               "[debug] [#{inspect(FakeClient)}] 200 " <>
+                 ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
+
+      assert logs =~ "[debug] [#{inspect(FakeClient)}] response body ok"
+    end
+
+    test "when the log level is given, logs the message with the given log level" do
+      logs = capture_log(fn -> FakeClient.get("/ok", opts: [log_level: :info]) end)
+
+      assert logs =~ "[info]  [#{inspect(Subject)}] GET /ok []"
 
       assert logs =~
                "[info]  [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[info]  [#{inspect(Subject)}] ok"
+      assert logs =~ "[info]  [#{inspect(Subject)}] response body ok"
     end
 
     test "when a filtered header is given, logs the message filtering the given headers" do
       logs =
         capture_log(fn ->
-          FakeClient.post("/ok", %{}, opts: [filter_headers: ["authorization", "content-type"]])
+          FakeClient.get("/ok", opts: [filter_headers: ["authorization", "content-type"]])
         end)
 
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] POST /ok [] [])
+      assert logs =~ "[debug] [#{inspect(Subject)}] GET /ok []"
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] 200 ) <>
+               "[debug] [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "[FILTERED]"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[debug] [#{inspect(Subject)}] ok"
+      assert logs =~ "[debug] [#{inspect(Subject)}] response body ok"
     end
 
-    test "when a filtered query param is given, logs the message filtering the given query param" do
+    test "when a filtered query param is given, logs the message filtering the given query params" do
       logs =
         capture_log(fn ->
-          FakeClient.post("/ok", %{},
+          FakeClient.get("/ok",
             query: [{:page, 1}, {:user, "test_user"}, {"password", "test password"}],
             opts: [filter_query_params: [:user, "password"]]
           )
         end)
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] POST /ok [{:page, 1}, {:user, "[FILTERED]"}, ) <>
-                 ~s({"password", "[FILTERED]"}] [])
+               "[debug] [#{inspect(Subject)}] " <>
+                 "GET /ok?page=1&user=[FILTERED]&password=[FILTERED] []"
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] 200 ) <>
+               "[debug] [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[debug] [#{inspect(Subject)}] ok"
+      assert logs =~ "[debug] [#{inspect(Subject)}] response body ok"
     end
 
-    test "when a filtered body is given, logs the message filtering the given body patterns" do
+    test "when a filtered body is given, logs the message filtering the given request body" do
       logs =
         capture_log(fn ->
-          FakeClient.post("/json", ~s({"password":"0123456789","somethingsafe":"ok"}),
+          FakeClient.post("/json", ~s({"password":"0123456789","some":"body"}),
             opts: [filter_body: [~r/"password":".*?"/]]
           )
         end)
 
-      assert logs =~
-               ~s([debug] [#{inspect(Subject)}] POST /json [] [])
-
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] {[FILTERED],"somethingsafe":"ok"})
+      assert logs =~ "[debug] [#{inspect(Subject)}] POST /json []"
+      assert logs =~ ~s([debug] [#{inspect(Subject)}] {[FILTERED],"some":"body"})
 
       assert logs =~
-               ~s([debug] [#{inspect(Subject)}] 200 ) <>
+               "[debug] [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
       assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"foo@bar.baz","response":"value"})
     end
 
-    test "when max entry length is given, " <>
-           "logs the request and response splitting the entries" do
-      body_prefix = String.duplicate("a", 80)
+    test "when the max entry length is given, " <>
+           "logs the request and the response splitting the body" do
+      request_body_slice1 = String.duplicate("x", 100)
+      request_body_slice2 = String.duplicate("y", 100)
 
       logs =
         capture_log(fn ->
-          FakeClient.post("/json", body_prefix <> "b", opts: [max_entry_length: 80])
+          FakeClient.post("/huge-response", request_body_slice1 <> request_body_slice2,
+            opts: [max_entry_length: 100]
+          )
         end)
 
-      assert logs =~
-               ~s([debug] [#{inspect(Subject)}] POST /json [] [])
-
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] #{body_prefix}\n)
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] b\n)
+      assert logs =~ "[debug] [#{inspect(Subject)}] POST /huge-response []"
+      assert logs =~ "[debug] [#{inspect(Subject)}] #{request_body_slice1}\n"
+      assert logs =~ "[debug] [#{inspect(Subject)}] #{request_body_slice2}\n"
 
       assert logs =~
                "[debug] [#{inspect(Subject)}] 200 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","response":"value"})
+      assert logs =~ "[debug] [#{inspect(Subject)}] #{String.duplicate("a", 100)}\n"
+      assert logs =~ "[debug] [#{inspect(Subject)}] #{String.duplicate("b", 100)}\n"
     end
 
     test "when response is an error, logs the response with error log level" do
       logs = capture_log(fn -> FakeClient.get("/error") end)
 
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] GET /error [] [])
+      assert logs =~ "[debug] [#{inspect(Subject)}] GET /error []"
 
       assert logs =~
                "[error] [#{inspect(Subject)}] 404 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[error] [#{inspect(Subject)}] error"
+      assert logs =~ "[error] [#{inspect(Subject)}] response body error"
     end
 
     test "when response is a redirect, logs response with warn log level" do
       logs = capture_log(fn -> FakeClient.get("/redirect", opts: [log_level: :info]) end)
 
-      assert logs =~ ~s([info]  [#{inspect(Subject)}] GET /redirect [] [])
+      assert logs =~ "[info]  [#{inspect(Subject)}] GET /redirect []"
 
       assert logs =~
                "[warn]  [#{inspect(Subject)}] 301 " <>
                  ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
 
-      assert logs =~ "[warn]  [#{inspect(Subject)}] moved"
+      assert logs =~ "[warn]  [#{inspect(Subject)}] response body moved"
     end
 
-    test "when it fails to connect, logs the error" do
+    test "when the request fails to connect, logs the error" do
       logs = capture_log(fn -> FakeClient.get("/connection-error") end)
 
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] GET /connection-error [] [])
+      assert logs =~ "[debug] [#{inspect(Subject)}] GET /connection-error []"
       assert logs =~ "[error] [#{inspect(Subject)}] :econnrefused"
-    end
-
-    test "when response body is a JSON, logs the message as a proper JSON" do
-      encoded_body = Jason.encode!(%{something: 1})
-
-      logs = capture_log(fn -> FakeClient.post("/json", encoded_body) end)
-
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] POST /json [] [])
-      assert logs =~ "[debug] [#{inspect(Subject)}] #{encoded_body}"
-
-      assert logs =~
-               "[debug] [#{inspect(Subject)}] 200 " <>
-                 ~s([{"content-type", "text/plain"}, {"authorization", "[FILTERED]"}])
-
-      assert logs =~ ~s([debug] [#{inspect(Subject)}] {"email":"[FILTERED]","response":"value"})
     end
   end
 end